Implement rating feature

+++ shared/src/commonMain/kotlin/com/example/moveeapp_compose_kmm/data/remote/model/movie/MovieDetailModel.kt
@@ -23,7 +23,7 @@ data class MovieDetailModel(
     @SerialName("production_companies") val productionCompanies: List<ProductionCompany>,
     @SerialName("production_countries") val productionCountries: List<ProductionCountry>,
     @SerialName("release_date") val releaseDate: String,
-    @SerialName("revenue") val revenue: Int,
+    @SerialName("revenue") val revenue: Long,
     @SerialName("runtime") val runtime: Int,
     @SerialName("spoken_languages") val spokenLanguages: List<SpokenLanguage>,
     @SerialName("status") val status: String,
